requirements.txt
torch
torchvision
numpy
Pillow
requests
matplotlib